### 2022-05-30 #4440 @Kevin-Mattheus-Moerman
@hechth can you fix those potentially missing DOIs? :point_up: You can call `@editorialbot check references` to check them again.
