+++ paper/paper.bib
@@ -3,6 +3,7 @@
   author={Yones, Sara and Csombordi, Rajmund and Komorowski, Jan and Diamanti, Klev},
   journal={bioRxiv},
   year={2021},
+  doi={10.1101/2021.02.28.433248},
   publisher={Cold Spring Harbor Laboratory}
 }
 
@@ -38,6 +39,7 @@ year = {2020}
 @article{tomasulo2002chemidplus,
   title={ChemIDplus-super source for chemical and drug information},
   author={Tomasulo, Patricia},
+  doi={10.1300/j115v21n01_04},
   journal={Medical reference services quarterly},
   volume={21},
   number={1},
@@ -54,6 +56,7 @@ year = {2020}
   number={D1},
   pages={D1388--D1395},
   year={2021},
+  doi={10.1093/nar/gkaa971},
   publisher={Oxford University Press}
 }
 
@@ -73,6 +76,7 @@ year = {2020}
   number={15},
   pages={1733--1734},
   year={2008},
+  doi={10.1093/bioinformatics/btn307},
   publisher={Oxford University Press}
 }
 
@@ -156,6 +160,7 @@ year = {2018}
   number={1},
   pages={1--19},
   year={2021},
+  doi={10.1186/s13321-021-00515-1},
   publisher={Springer}
 }
 
