### 2022-05-31 a468045
rev and minimize the codeql config
