### 2022-06-19 #4440 @csoneson
👋🏻 @smsaladi, @marshallmcdonnell, @chryswoods - would you be interested in reviewing this submission for JOSS?

> MSMetaEnhancer: A Python package for mass spectra metadata annotation
