### 2022-06-20 #4440 @smsaladi
Thanks for reaching out to me. Unfortunately, I cannot assist with
reviewing this manuscript right now.

Shyam

On Sun, Jun 19, 2022 at 4:26 PM Charlotte Soneson ***@***.***>
wrote:

> 👋🏻 @smsaladi <https://github.com/smsaladi>, @marshallmcdonnell
> <https://github.com/marshallmcdonnell>, @chryswoods
> <https://github.com/chryswoods> - would you be interested in reviewing
> this submission for JOSS?
>
> MSMetaEnhancer: A Python package for mass spectra metadata annotation
>
