### 2022-06-21 RECETOX/MSMetaEnhancer#111 @hechth
Fix DOIs
```

OK DOIs

- None

MISSING DOIs

- None

INVALID DOIs

- 10.1101/2021.02.28.433248 is INVALID
- 10.21105/joss.02411 is INVALID
- 10.1093/bioinformatics/btq476 is INVALID
- 10.1300/j115v21n01_04 is INVALID
- 10.1093/nar/gkaa971 is INVALID
- 10.1093/bioinformatics/btn307 is INVALID
- 10.1093/nar/gky379 is INVALID
- 10.1038/s41592-018-0046-7 is INVALID
- 10.1186/s13321-021-00515-1 is INVALID
- 10.5281/zenodo.6035335 is INVALID
- 10.1093/nar/gkab1062 is INVALID
- 10.1186/1471-2105-11-5 is INVALID
- 10.3390/metabo8010016 is INVALID
- 10.3897/rio.8.e83031 is INVALID
- 10.3389/fpubh.2021.622558 is INVALID
- 10.1007/s13361-016-1589-4 is INVALID
- 10.1016/S1044-0305(98)00159-7 is INVALID
- 10.1002/jms.3131 is INVALID
```

_Originally posted by @editorialbot in https://github.com/openjournals/joss-reviews/issues/4494#issuecomment-1161322273_
