### 2022-06-21 RECETOX/MSMetaEnhancer#111 @martenson
I checked some DOIs and they seem fine, do they want some other format?
