### 2022-06-21 #4494 @hechth
@csoneson Is there maybe some issue with the DOI checker? We had a look at the DOIs and they actually seem fine.
