### 2022-06-21 #4494 @csoneson
Hm, not sure what is going on here. @openjournals/dev - any idea? The DOIs were considered valid in the pre-review issue.
