### 2022-06-27 RECETOX/MSMetaEnhancer#111 @xtrojak
Seems like there was just an issue with the validator ([here](https://github.com/openjournals/joss-reviews/issues/4494#issuecomment-1161431397) it worked again).
