### 2022-07-05 RECETOX/MSMetaEnhancer#112 @hechth
Hi @chryswoods! Thank you very much for your comment!

This is actually something that is related to not updating our tests after a new release of matchms which fixed the SMILES validation (https://github.com/matchms/matchms/pull/337).

We will update our tests and the dependency on matchms to be consistent!
