### 2022-07-05 RECETOX/MSMetaEnhancer#113 @marshallmcdonnell
[Exploratory] Using the circuit breaker design pattern for web calls
This is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/4494)
NOTE: Only an optional suggestion, definitely not a requested change for the review

**Description**
Looking over [`MSMetaEnhancer.libs.converters.web.WebConverter.loop_request`](https://github.com/RECETOX/MSMetaEnhancer/blob/main/MSMetaEnhancer/libs/converters/web/WebConverter.py#L67-L91), this "smells" like the [Circuit Breaker Design Pattern](https://en.wikipedia.org/wiki/Circuit_breaker_design_pattern)

I was just going to bring this to your attention to see if it might be of value to adhere to the pattern since it is very close already.

Also, there are multiple Python packages available to add this capability to functions or methods:
 * [circuitbreaker](https://github.com/fabfuel/circuitbreaker)
 * [pybreaker](https://github.com/danielfm/pybreaker)
 * [python-circuit](https://github.com/edgeware/python-circuit)

And if it helps, I investigated these three myself in the past for a very similar use case:
https://github.com/marshallmcdonnell/python-circuit-breaker-examples

Overall, just hoping this helps and feel free to close if not helpful.
