### 2022-07-05 RECETOX/MSMetaEnhancer#114 @marshallmcdonnell
[Docs] Restructure documentation ordering for ReadTheDocs
This is part of my https://github.com/openjournals/joss-reviews/issues/4494
NOTE: Only an optional suggestion, definitely not a requested change for the review

**Description**

One thing I noticed on the MSMetaEnhancer ReadTheDocs site is that the ["introduction"](https://msmetaenhancer.readthedocs.io/en/latest/readme.html) comes after the [Python API documentation](https://msmetaenhancer.readthedocs.io/en/latest/MSMetaEnhancer.html)

I don't think you have to change this since it is a style thing.

But I think for the reader, it would be better to have them first see the intro doc to understand the project before they dive into the API.

I believe that ordering can be set easily in [this file](https://github.com/RECETOX/MSMetaEnhancer/blob/main/docs/source/index.rst), if you would like to change it.

Overall, just hoping this helps and feel free to close if not helpful.
