### 2022-07-05 #4494 @marshallmcdonnell
@hechth, I'm excited about your software package, big fan of tools that help add and standardize metadata! Clean code base, was very readable. Good intuitive design for the Python API.  Also tried out the UMSA Galaxy tool, worked great!

I honestly just have non-mandatory suggestions. No major issues from me so I don't see any issues publishing in JOSS.

**Suggestions**
- https://github.com/RECETOX/MSMetaEnhancer/issues/113
- https://github.com/RECETOX/MSMetaEnhancer/issues/114
- https://github.com/RECETOX/MSMetaEnhancer/issues/115

Tests ran as expected.
Installation via conda was as expected. Able to run the example in README.
USMA Galaxy install ran as expected.

Let me know if you have questions about these issues and happy to engage in a discussion.
