### 2022-07-07 #4494 @chryswoods
@hechth - I agree with @marshallmcdonnell that this is a great tool. Very clean code.

I agree with the suggestion of updating the presentation of the documentation on the readthedocs website. It can help to attract new users if they can see a general introduction plus some examples first. The API docs are useful, but tend to be something that someone will read after they have used the tool for longer, and progressed beyond just basic usage. In my experience, you can never have enough "basic usage" tutorials ;-).

I am stuck with my review by the test failure and also a runtime error when running the example. I've reported these to the repo.

* https://github.com/RECETOX/MSMetaEnhancer/issues/112#issue-1293189680
* https://github.com/RECETOX/MSMetaEnhancer/issues/116#issue-1297704537

I did struggle to install the software from conda. This was mostly because of unresolvable dependencies, even in a vanilla miniconda environment. I ended up getting it working by using a vanilla minimamba on Linux.

I agree with @marshallmcdonnell it may make packaging easier if you unified the environment and package management. I noticed that your `requirements.txt` has lots of pinning of package versions, while your conda recipe was relatively unpinned. I recommend removing the pins from `requirements.txt` or making it match up more with what is in the conda recipe.
