### 2022-07-11 RECETOX/MSMetaEnhancer#116 @xtrojak
Thank you for pointing this out @chryswoods. The `Application.annotate_spectra` is actually meant to be run only once, so the code in README was misleading. I removed the code in #117, where you can also find more details.
