### 2022-07-11 RECETOX/MSMetaEnhancer#112 @xtrojak
You are right @hechth, `matchms` version 0.16.0 employed a new approach for checking SMILES. I removed the obsolete test (`Curator` is not supposed to remove SMILES `'C'` anymore) in #118. Thank you @chryswoods!
