### 2022-07-22 4b6481a
Simplified yaml files
