+++ conda/environment-dev.yml
@@ -4,7 +4,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python >=3.7,<3.9
+  - python >=3.7
   - matchms >=0.14.0
   - pandas
   - scipy
