+++ conda/environment-build.yml
@@ -4,4 +4,4 @@ channels:
 dependencies:
   - conda-build
   - conda-verify
-  - python >=3.7,<3.9
+  - python >=3.7
