### 2022-08-05 c445b63
Simplify requirements.txt
