+++ requirements.txt
@@ -1,17 +1,18 @@
 matchms~=0.14.0
-pandas~=1.2.4
-requests~=2.25.1
-mock~=4.0.3
-aiohttp~=3.7.4.post0
-asyncstdlib~=3.9.2
-frozendict~=2.0.3
+pandas
+scipy
+requests
+aiohttp
+asyncstdlib
+frozendict
+twine
+mock
 pytest
 pytest-aiohttp
 pytest-cov
 pytest-dependency
-tabulate~=0.8.9
-sphinx==4.2.0
-sphinx_rtd_theme==1.0.0
-myst-parser==0.15.2
-rdkit-pypi~=2021.9.3
-multidict~=5.2.0
+tabulate
+sphinx
+sphinx_rtd_theme
+rdkit-pypi
+multidict
