### 2022-08-05 c181810
Simplify environment-dev.yml using requirements.txt
