### 2022-08-05 d905123
Separated dev and run requirements, updated that in setup.py
