+++ conda/environment-dev.yml
@@ -1,10 +1,9 @@
 name: MSMetaEnhancer-dev
 channels:
-  - bioconda
-  - conda-forge
   - defaults
 dependencies:
   - python >=3.7
   - pip
   - pip:
     - -r ../requirements.txt
+    - -r ../requirements-dev.txt
