### 2022-08-08 ac3a3bf
Changed order of readme and docs in index file
