+++ docs/source/index.rst
@@ -9,8 +9,8 @@ Welcome to MSMetaEnhancer's documentation!
 .. toctree::
    :maxdepth: 1
 
-   Documentation <MSMetaEnhancer>
    Readme <readme>
+   Documentation <MSMetaEnhancer>
    CHANGELOG
 
 Indices and tables
