### 2022-08-08 fd37e3c
Add conda build workflow
