### 2022-08-08 0e53510
Update meta.yaml with pip source
