### 2022-08-09 df42dbb
revert meta.yaml, include req. files to manifest
