### 2022-08-09 RECETOX/MSMetaEnhancer#115 @xtrojak
Thank you for suggestions @marshallmcdonnell.

Unfortunately, it is not possible to unify the environment and packaging management completely, conda and pypi package managers are just too separated (see details in this [discussion](https://github.com/conda/conda-build/issues/548)).

Still we were able to improve the local environment and requirements files as you suggested, so thank you for that!

---

> Yet, I don't see the conda package publishing in CD. Maybe I am missing something?

Conda package publishing is done automatically by a [bioconda bot](https://github.com/BiocondaBot) which checks for releases in out repository and creates an [update PR](
https://github.com/bioconda/bioconda-recipes/pulls?q=is%3Apr+msmetaenhancer+is%3Aclosed) in [bioconda recipes repository](https://github.com/bioconda/bioconda-recipes).
