### 2022-08-17 RECETOX/MSMetaEnhancer#113 @maximskorik
https://github.com/sgenya/aiocircuitbreaker/issues/4
