### 2022-08-18 d154c4f
make `FAILURE_THRESHOLD` a class attribute
