### 2022-08-18 bf49b9e
add `aiocircuitbreaker` to requirements.txt & setup.py
