+++ requirements.txt
@@ -15,3 +15,4 @@ sphinx_rtd_theme==1.0.0
 myst-parser==0.15.2
 rdkit-pypi~=2021.9.3
 multidict~=5.2.0
+aiocircuitbreaker~=2.0.0
