+++ MSMetaEnhancer/libs/utils/Errors.py
@@ -20,8 +20,9 @@ class SourceAttributeNotAvailable(Exception):
 
 class ServiceNotAvailable(Exception):
     @staticmethod
-    async def async_raise(*args, **kwargs):
-        raise ServiceNotAvailable('Service not available.')
+    async def raise_circuitbreaker(*args):
+        converter_name = args[0].converter_name
+        raise ServiceNotAvailable(f'Service {converter_name} not available.')
 
 
 class UnknownResponse(Exception):
