### 2022-08-18 a77adec
combine fixtures
