+++ conda/environment-dev.yml
@@ -2,7 +2,7 @@ name: MSMetaEnhancer-dev
 channels:
   - defaults
 dependencies:
-  - python >=3.7
+  - python >=3.9
   - pip
   - pip:
     - -r ../requirements.txt
