+++ .readthedocs.yaml
@@ -5,13 +5,18 @@
 # Required
 version: 2
 
+# Set the version of Python and other tools you might need
+build:
+  os: ubuntu-20.04
+  tools:
+    python: "3.9"
+
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
   builder: html
   configuration: docs/source/conf.py
 
 python:
-  version: "3.8"
   install:
     - requirements: requirements.txt
     - requirements: requirements-dev.txt
