### 2022-08-26 f1aec5e
Update python version in CI
