+++ .github/workflows/python-package.yml
@@ -16,7 +16,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
-        python-version: [3.8, 3.9]
+        python-version: [3.9, 3.10]
 
     steps:
     - uses: actions/checkout@v2
