### 2022-09-12 #4494 @hechth
@csoneson Sorry for the delay and absence - I think we fixed all issues that were mentioned in the review, I am now testing the newly deployed version on our Galaxy instance to make sure everything runs also smoothly in production. I will make a release on Zenodo with that version afterwards so we can continue with the publication process :).

Thank you very much for your time and patience.
