### 2022-09-13 #4494 @hechth
@csoneson @chryswoods and @marshallmcdonnell - Version 0.2.4 is released and also available on Galaxy. It contains a simplified setup, and the circuit breaker pattern for aiohttp.

Please let me know whether the paper can be considered as accepted so that I can create the repository on Zenodo :).
