### 2022-09-14 #4494 @chryswoods
Thanks - I also really appreciate all the work you have done. I am though still struggling to install the software following the instructions on your website. I have tried on a variety of platforms. For example, it doesn't install on Linux using a new, clean installation of the latest version of miniconda. Here's all the output;

```
(base) [opc@metawards ~]$ conda create --name MSMetaEnhancer python=3.9
Collecting package metadata (current_repodata.json): done
Solving environment: done

## Package Plan ##

  environment location: /home/opc/miniconda3/envs/MSMetaEnhancer

  added / updated specs:
    - python=3.9

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    _libgcc_mutex-0.1          |      conda_forge           3 KB  conda-forge
    _openmp_mutex-4.5          |            2_gnu          23 KB  conda-forge
    bzip2-1.0.8                |       h7f98852_4         484 KB  conda-forge
    ca-certificates-2022.6.15.2|       ha878542_0         150 KB  conda-forge
    ld_impl_linux-64-2.36.1    |       hea4e1c9_2         667 KB  conda-forge
    libffi-3.4.2               |       h7f98852_5          57 KB  conda-forge
    libgcc-ng-12.1.0           |      h8d9b700_16         940 KB  conda-forge
    libgomp-12.1.0             |      h8d9b700_16         459 KB  conda-forge
    libnsl-2.0.0               |       h7f98852_0          31 KB  conda-forge
    libsqlite-3.39.3           |       h753d276_0         789 KB  conda-forge
    libuuid-2.32.1             |    h7f98852_1000          28 KB  conda-forge
    libzlib-1.2.12             |       h166bdaf_3          65 KB  conda-forge
    ncurses-6.3                |       h27087fc_1        1002 KB  conda-forge
    openssl-3.0.5              |       h166bdaf_2         2.8 MB  conda-forge
    pip-22.2.2                 |     pyhd8ed1ab_0         1.5 MB  conda-forge
    python-3.9.13              |h2660328_0_cpython        26.7 MB  conda-forge
    readline-8.1.2             |       h0f457ee_0         291 KB  conda-forge
    setuptools-65.3.0          |     pyhd8ed1ab_1         782 KB  conda-forge
    sqlite-3.39.3              |       h4ff8645_0         789 KB  conda-forge
    tk-8.6.12                  |       h27826a3_0         3.3 MB  conda-forge
    tzdata-2022c               |       h191b570_0         119 KB  conda-forge
    wheel-0.37.1               |     pyhd8ed1ab_0          31 KB  conda-forge
    xz-5.2.6                   |       h166bdaf_0         409 KB  conda-forge
    ------------------------------------------------------------
                                           Total:        41.2 MB

The following NEW packages will be INSTALLED:

  _libgcc_mutex      conda-forge/linux-64::_libgcc_mutex-0.1-conda_forge
  _openmp_mutex      conda-forge/linux-64::_openmp_mutex-4.5-2_gnu
  bzip2              conda-forge/linux-64::bzip2-1.0.8-h7f98852_4
  ca-certificates    conda-forge/linux-64::ca-certificates-2022.6.15.2-ha878542_0
  ld_impl_linux-64   conda-forge/linux-64::ld_impl_linux-64-2.36.1-hea4e1c9_2
  libffi             conda-forge/linux-64::libffi-3.4.2-h7f98852_5
  libgcc-ng          conda-forge/linux-64::libgcc-ng-12.1.0-h8d9b700_16
  libgomp            conda-forge/linux-64::libgomp-12.1.0-h8d9b700_16
  libnsl             conda-forge/linux-64::libnsl-2.0.0-h7f98852_0
  libsqlite          conda-forge/linux-64::libsqlite-3.39.3-h753d276_0
  libuuid            conda-forge/linux-64::libuuid-2.32.1-h7f98852_1000
  libzlib            conda-forge/linux-64::libzlib-1.2.12-h166bdaf_3
  ncurses            conda-forge/linux-64::ncurses-6.3-h27087fc_1
  openssl            conda-forge/linux-64::openssl-3.0.5-h166bdaf_2
  pip                conda-forge/noarch::pip-22.2.2-pyhd8ed1ab_0
  python             conda-forge/linux-64::python-3.9.13-h2660328_0_cpython
  readline           conda-forge/linux-64::readline-8.1.2-h0f457ee_0
  setuptools         conda-forge/noarch::setuptools-65.3.0-pyhd8ed1ab_1
  sqlite             conda-forge/linux-64::sqlite-3.39.3-h4ff8645_0
  tk                 conda-forge/linux-64::tk-8.6.12-h27826a3_0
  tzdata             conda-forge/noarch::tzdata-2022c-h191b570_0
  wheel              conda-forge/noarch::wheel-0.37.1-pyhd8ed1ab_0
  xz                 conda-forge/linux-64::xz-5.2.6-h166bdaf_0

Proceed ([y]/n)? y

Downloading and Extracting Packages
bzip2-1.0.8          | 484 KB    | ################################################################################################################# | 100%
ca-certificates-2022 | 150 KB    | ################################################################################################################# | 100%
setuptools-65.3.0    | 782 KB    | ################################################################################################################# | 100%
python-3.9.13        | 26.7 MB   | ################################################################################################################# | 100%
xz-5.2.6             | 409 KB    | ################################################################################################################# | 100%
tzdata-2022c         | 119 KB    | ################################################################################################################# | 100%
_openmp_mutex-4.5    | 23 KB     | ################################################################################################################# | 100%
wheel-0.37.1         | 31 KB     | ################################################################################################################# | 100%
_libgcc_mutex-0.1    | 3 KB      | ################################################################################################################# | 100%
libgcc-ng-12.1.0     | 940 KB    | ################################################################################################################# | 100%
libnsl-2.0.0         | 31 KB     | ################################################################################################################# | 100%
libgomp-12.1.0       | 459 KB    | ################################################################################################################# | 100%
readline-8.1.2       | 291 KB    | ################################################################################################################# | 100%
openssl-3.0.5        | 2.8 MB    | ################################################################################################################# | 100%
libffi-3.4.2         | 57 KB     | ################################################################################################################# | 100%
libuuid-2.32.1       | 28 KB     | ################################################################################################################# | 100%
ncurses-6.3          | 1002 KB   | ################################################################################################################# | 100%
libzlib-1.2.12       | 65 KB     | ################################################################################################################# | 100%
ld_impl_linux-64-2.3 | 667 KB    | ################################################################################################################# | 100%
pip-22.2.2           | 1.5 MB    | ################################################################################################################# | 100%
sqlite-3.39.3        | 789 KB    | ################################################################################################################# | 100%
tk-8.6.12            | 3.3 MB    | ################################################################################################################# | 100%
libsqlite-3.39.3     | 789 KB    | ################################################################################################################# | 100%
Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate MSMetaEnhancer
#
# To deactivate an active environment, use
#
#     $ conda deactivate

(base) [opc@metawards ~]$ conda activate MSMetaEnhancer
(MSMetaEnhancer) [opc@metawards ~]$ conda install --channel bioconda --channel conda-forge MSMetaEnhancer
Collecting package metadata (current_repodata.json): done
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: -
Found conflicts! Looking for incompatible packages.
This can take several minutes.  Press CTRL-C to abort.
failed

UnsatisfiableError: The following specifications were found to be incompatible with each other:

Output in format: Requested package -> Available versions

Package tk conflicts for:
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> tk[version='8.5.*|8.6.*|>=8.6.10,<8.7.0a0|>=8.6.11,<8.7.0a0|>=8.6.12,<8.7.0a0|>=8.6.9,<8.7.0a0|>=8.6.8,<8.7.0a0|>=8.6.7,<8.7.0a0']
python=3.9 -> tk[version='>=8.6.10,<8.7.0a0|>=8.6.11,<8.7.0a0|>=8.6.12,<8.7.0a0']
setuptools -> python[version='>=3.8'] -> tk[version='8.5.*|8.6.*|>=8.6.10,<8.7.0a0|>=8.6.11,<8.7.0a0|>=8.6.12,<8.7.0a0|>=8.6.9,<8.7.0a0|>=8.6.8,<8.7.0a0|>=8.6.7,<8.7.0a0']
pip -> python[version='>=3.7'] -> tk[version='8.5.*|8.6.*|>=8.6.10,<8.7.0a0|>=8.6.11,<8.7.0a0|>=8.6.12,<8.7.0a0|>=8.6.9,<8.7.0a0|>=8.6.8,<8.7.0a0|>=8.6.7,<8.7.0a0']
msmetaenhancer -> python[version='>=3.9'] -> tk[version='>=8.6.10,<8.7.0a0|>=8.6.11,<8.7.0a0|>=8.6.12,<8.7.0a0|>=8.6.9,<8.7.0a0|>=8.6.8,<8.7.0a0|>=8.6.7,<8.7.0a0']
tk

Package pip conflicts for:
setuptools -> python[version='>=3.8'] -> pip
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> pip
msmetaenhancer -> python[version='>=3.9'] -> pip
python=3.9 -> pip
pip

Package pypy3.7 conflicts for:
setuptools -> python[version='>=3.7,<3.8.0a0'] -> pypy3.7[version='7.3.*|7.3.3.*|7.3.4.*|7.3.5.*|7.3.7.*|>=7.3.4']
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> pypy3.7[version='7.3.3.*|7.3.4.*|7.3.5.*|7.3.7.*|>=7.3.7|>=7.3.5|>=7.3.3']
setuptools -> pypy3.7[version='>=7.3.3|>=7.3.5|>=7.3.7']
msmetaenhancer -> aiohttp -> pypy3.7[version='7.3.3.*|7.3.4.*|7.3.5.*|7.3.7.*|>=7.3.3|>=7.3.4|>=7.3.7|>=7.3.5']
pip -> python[version='>=3.7'] -> pypy3.7[version='7.3.3.*|7.3.4.*|7.3.5.*|7.3.7.*|>=7.3.7|>=7.3.5|>=7.3.3']

Package _libgcc_mutex conflicts for:
libnsl -> libgcc-ng[version='>=9.4.0'] -> _libgcc_mutex==0.1=conda_forge
xz -> libgcc-ng[version='>=12'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
python=3.9 -> libgcc-ng[version='>=12'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
libffi -> libgcc-ng[version='>=9.4.0'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
tk -> libgcc-ng[version='>=9.4.0'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
libsqlite -> libgcc-ng[version='>=12'] -> _libgcc_mutex==0.1=conda_forge
sqlite -> libgcc-ng[version='>=12'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
libgomp -> _libgcc_mutex==0.1=conda_forge
libgcc-ng -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
_openmp_mutex -> _libgcc_mutex==0.1=conda_forge
ncurses -> libgcc-ng[version='>=10.3.0'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
libzlib -> libgcc-ng[version='>=12'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
readline -> libgcc-ng[version='>=12'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
bzip2 -> libgcc-ng[version='>=9.3.0'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']
_libgcc_mutex
libuuid -> libgcc-ng[version='>=9.3.0'] -> _libgcc_mutex[version='*|0.1',build='main|conda_forge']

Package zlib conflicts for:
python=3.9 -> zlib[version='>=1.2.11,<1.3.0a0']
setuptools -> python[version='>=3.8'] -> zlib[version='1.2.*|1.2.11|>=1.2.11,<1.3.0a0|>=1.2.12,<1.3.0a0|1.2.8|1.2.11.*']
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> zlib[version='1.2.*|1.2.11|>=1.2.11,<1.3.0a0|1.2.8|1.2.11.*']
pip -> python[version='>=3.7'] -> zlib[version='1.2.*|1.2.11|>=1.2.11,<1.3.0a0|1.2.8|1.2.11.*']
sqlite -> zlib[version='>=1.2.11,<1.3.0a0|>=1.2.12,<1.3.0a0']
tk -> zlib[version='>=1.2.11,<1.3.0a0']
msmetaenhancer -> python[version='>=3.9'] -> zlib[version='>=1.2.11,<1.3.0a0']
python=3.9 -> sqlite[version='>=3.38.5,<4.0a0'] -> zlib[version='>=1.2.12,<1.3.0a0']

Package readline conflicts for:
setuptools -> python[version='>=3.8'] -> readline[version='6.2.*|7.0|>=7.0,<8.0a0|>=8.0,<9.0a0|>=8.1,<9.0a0|>=8.1.2,<9.0a0|7.0.*']
pip -> python[version='>=3.7'] -> readline[version='6.2.*|7.0|>=7.0,<8.0a0|>=8.0,<9.0a0|>=8.1,<9.0a0|>=8.1.2,<9.0a0|7.0.*']
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> readline[version='6.2.*|7.0|>=7.0,<8.0a0|>=8.0,<9.0a0|>=8.1,<9.0a0|>=8.1.2,<9.0a0|7.0.*']
python=3.9 -> readline[version='>=8.0,<9.0a0|>=8.1,<9.0a0']
sqlite -> readline[version='6.2.*|7.0|7.0.*|>=7.0,<8.0a0|>=8.0,<9.0a0|>=8.1,<9.0a0|>=8.1.2,<9.0a0']
python=3.9 -> sqlite[version='>=3.38.5,<4.0a0'] -> readline[version='>=8.1.2,<9.0a0']
msmetaenhancer -> python[version='>=3.9'] -> readline[version='>=7.0,<8.0a0|>=8.0,<9.0a0|>=8.1,<9.0a0|>=8.1.2,<9.0a0']
readline

Package _openmp_mutex conflicts for:
xz -> libgcc-ng[version='>=12'] -> _openmp_mutex[version='>=4.5']
ncurses -> libgcc-ng[version='>=10.3.0'] -> _openmp_mutex[version='>=4.5']
libffi -> libgcc-ng[version='>=9.4.0'] -> _openmp_mutex[version='>=4.5']
tk -> libgcc-ng[version='>=9.4.0'] -> _openmp_mutex[version='>=4.5']
libsqlite -> libgcc-ng[version='>=12'] -> _openmp_mutex[version='>=4.5']
python=3.9 -> libgcc-ng[version='>=12'] -> _openmp_mutex[version='>=4.5']
libzlib -> libgcc-ng[version='>=12'] -> _openmp_mutex[version='>=4.5']
readline -> libgcc-ng[version='>=12'] -> _openmp_mutex[version='>=4.5']
sqlite -> libgcc-ng[version='>=12'] -> _openmp_mutex[version='>=4.5']
libnsl -> libgcc-ng[version='>=9.4.0'] -> _openmp_mutex[version='>=4.5']
libuuid -> libgcc-ng[version='>=9.3.0'] -> _openmp_mutex[version='>=4.5']
libgcc-ng -> _openmp_mutex[version='>=4.5']
_openmp_mutex
bzip2 -> libgcc-ng[version='>=9.3.0'] -> _openmp_mutex[version='>=4.5']

Package libffi conflicts for:
libffi
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> libffi[version='>=3.2.1,<3.3.0a0|>=3.3,<3.4.0a0|>=3.4.2,<3.5.0a0']
python=3.9 -> libffi[version='>=3.2.1,<3.3.0a0|>=3.3,<3.4.0a0|>=3.4.2,<3.5.0a0']
setuptools -> python[version='>=3.8'] -> libffi[version='>=3.2.1,<3.3.0a0|>=3.3,<3.4.0a0|>=3.4.2,<3.5.0a0']
msmetaenhancer -> python[version='>=3.9'] -> libffi[version='>=3.2.1,<3.3.0a0|>=3.3,<3.4.0a0|>=3.4.2,<3.5.0a0']
pip -> python[version='>=3.7'] -> libffi[version='>=3.2.1,<3.3.0a0|>=3.3,<3.4.0a0|>=3.4.2,<3.5.0a0']

Package libgcc-ng conflicts for:
sqlite -> libgcc-ng[version='>=10.3.0|>=12|>=9.4.0|>=9.3.0|>=7.5.0|>=7.3.0|>=4.9']
setuptools -> python[version='>=3.8'] -> libgcc-ng[version='>=10.3.0|>=12|>=9.4.0|>=9.3.0|>=7.5.0|>=7.3.0|>=4.9']
bzip2 -> libgcc-ng[version='>=4.9|>=7.3.0|>=7.5.0|>=9.3.0']
libffi -> libgcc-ng[version='>=4.9|>=7.3.0|>=7.5.0|>=9.4.0']
tk -> libzlib[version='>=1.2.11,<1.3.0a0'] -> libgcc-ng[version='>=10.3.0|>=12']
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> libgcc-ng[version='>=10.3.0|>=12|>=9.4.0|>=9.3.0|>=7.5.0|>=7.3.0|>=4.9']
python=3.9 -> bzip2[version='>=1.0.8,<2.0a0'] -> libgcc-ng[version='>=4.9|>=7.3.0']
msmetaenhancer -> aiohttp -> libgcc-ng[version='>=10.3.0|>=9.4.0|>=9.3.0|>=7.5.0|>=7.3.0|>=4.9|>=7.2.0|>=12']
readline -> libgcc-ng[version='>=12|>=9.3.0|>=7.5.0|>=7.3.0|>=4.9']
python=3.9 -> libgcc-ng[version='>=10.3.0|>=12|>=9.4.0|>=9.3.0|>=7.5.0']
pip -> python[version='>=3.7'] -> libgcc-ng[version='>=10.3.0|>=12|>=9.4.0|>=9.3.0|>=7.5.0|>=7.3.0|>=4.9']
libnsl -> libgcc-ng[version='>=9.4.0']
readline -> ncurses[version='>=6.3,<7.0a0'] -> libgcc-ng[version='>=10.3.0|>=9.4.0']
tk -> libgcc-ng[version='>=4.9|>=7.3.0|>=7.5.0|>=9.4.0']
libgcc-ng
libsqlite -> libgcc-ng[version='>=12']
libuuid -> libgcc-ng[version='>=7.3.0|>=9.3.0']
xz -> libgcc-ng[version='>=12|>=7.5.0|>=7.3.0|>=4.9']
_openmp_mutex -> llvm-openmp[version='>=9.0.1'] -> libgcc-ng[version='>=7.3.0']
libzlib -> libgcc-ng[version='>=10.3.0|>=12|>=7.5.0']
ncurses -> libgcc-ng[version='>=10.3.0|>=9.4.0|>=7.5.0|>=7.3.0|>=4.9']

Package ncurses conflicts for:
pip -> python[version='>=3.7'] -> ncurses[version='5.9.*|5.9|>=6.1,<7.0.0a0|>=6.2,<6.3.0a0|>=6.3,<7.0a0|>=6.2,<7.0.0a0']
setuptools -> python[version='>=3.8'] -> ncurses[version='5.9.*|5.9|>=6.1,<7.0.0a0|>=6.2,<6.3.0a0|>=6.3,<7.0a0|>=6.2,<7.0.0a0']
python=3.9 -> ncurses[version='>=6.2,<7.0.0a0|>=6.3,<7.0a0']
msmetaenhancer -> python[version='>=3.9'] -> ncurses[version='>=6.1,<7.0.0a0|>=6.2,<6.3.0a0|>=6.3,<7.0a0|>=6.2,<7.0.0a0']
readline -> ncurses[version='5.9.*|>=6.1,<7.0.0a0|>=6.2,<7.0.0a0|>=6.3,<7.0a0']
python=3.9 -> readline[version='>=8.0,<9.0a0'] -> ncurses[version='>=6.1,<7.0.0a0']
sqlite -> ncurses[version='5.9|5.9.*|>=6.1,<7.0.0a0|>=6.2,<7.0.0a0|>=6.3,<7.0a0']
ncurses
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> ncurses[version='5.9.*|5.9|>=6.1,<7.0.0a0|>=6.2,<6.3.0a0|>=6.3,<7.0a0|>=6.2,<7.0.0a0']

Package packaging conflicts for:
wheel -> packaging[version='>=20.2']
pip -> wheel -> packaging[version='>=20.2']

Package libgomp conflicts for:
libgomp
libgcc-ng -> _openmp_mutex[version='>=4.5'] -> libgomp[version='>=7.3.0|>=7.5.0']
_openmp_mutex -> libgomp[version='>=7.3.0|>=7.5.0']

Package llvm-openmp conflicts for:
libgcc-ng -> _openmp_mutex[version='>=4.5'] -> llvm-openmp[version='>=9.0.1']
_openmp_mutex -> llvm-openmp[version='>=9.0.1']

Package libstdcxx-ng conflicts for:
setuptools -> python[version='>=3.8'] -> libstdcxx-ng[version='>=4.9|>=7.3.0|>=7.5.0|>=9.3.0|>=9.4.0']
python=3.9 -> libstdcxx-ng[version='>=7.5.0|>=9.3.0']
sqlite -> libstdcxx-ng[version='>=4.9|>=7.3.0']
sqlite -> ncurses[version='>=6.2,<7.0.0a0'] -> libstdcxx-ng[version='>=7.5.0']
ncurses -> libstdcxx-ng[version='>=4.9|>=7.3.0|>=7.5.0']
pip -> python[version='>=3.7'] -> libstdcxx-ng[version='>=4.9|>=7.3.0|>=7.5.0|>=9.3.0|>=9.4.0']
python=3.9 -> libffi[version='>=3.4.2,<3.5.0a0'] -> libstdcxx-ng[version='>=4.9|>=7.3.0|>=9.4.0']
libffi -> libstdcxx-ng[version='>=4.9|>=7.3.0|>=7.5.0|>=9.4.0']
wheel -> python[version='!=3.0,!=3.1,!=3.2,!=3.3,!=3.4'] -> libstdcxx-ng[version='>=4.9|>=7.3.0|>=7.5.0|>=9.3.0|>=9.4.0']
msmetaenhancer -> pandas -> libstdcxx-ng[version='>=10.3.0|>=12|>=9.4.0|>=9.3.0|>=7.5.0|>=7.3.0|>=4.9']
_openmp_mutex -> llvm-openmp[version='>=9.0.1'] -> libstdcxx-ng[version='>=7.3.0']
readline -> ncurses[version='>=6.2,<7.0.0a0'] -> libstdcxx-ng[version='>=4.9|>=7.3.0|>=7.5.0']

Package sqlite conflicts for:
python=3.9 -> sqlite[version='>=3.33.0,<4.0a0|>=3.34.0,<4.0a0|>=3.35.5,<4.0a0|>=3.36.0,<4.0a0|>=3.37.0,<4.0a0|>=3.37.1,<4.0a0|>=3.38.5,<4.0a0']
python=3.9 -> pypy3.9=7.3.9 -> sqlite[version='>=3.38.2,<4.0a0|>=3.39.1,<4.0a0|>=3.39.2,<4.0a0']
msmetaenhancer -> python[version='>=3.9'] -> sqlite[version='>=3.24.0,<4.0a0|>=3.25.1,<4.0a0|>=3.25.2,<4.0a0|>=3.25.3,<4.0a0|>=3.26.0,<4.0a0|>=3.28.0,<4.0a0|>=3.30.1,<4.0a0|>=3.32.3,<4.0a0|>=3.33.0,<4.0a0|>=3.34.0,<4.0a0|>=3.35.5,<4.0a0|>=3.36.0,<4.0a0|>=3.37.0,<4.0a0|>=3.37.1,<4.0a0|>=3.38.5,<4.0a0']
pip -> python[version='>=3.7'] -> sqlite[version='3.13.*|3.20.*|>=3.24.0,<4.0a0|>=3.25.1,<4.0a0|>=3.25.2,<4.0a0|>=3.25.3,<4.0a0|>=3.26.0,<4.0a0|>=3.28.0,<4.0a0|>=3.30.1,<4.0a0|>=3.32.3,<4.0a0|>=3.33.0,<4.0a0|>=3.34.0,<4.0a0|>=3.35.5,<4.0a0|>=3.36.0,<4.0a0|>=3.37.0,<4.0a0|>=3.37.1,<4.0a0|>=3.38.5,<4.0a0|3.9.*']
[…]
