### 2022-09-14 #4494 @chryswoods
Ok - I've fixed it. The conda channel order priority was borked on my computer. I removed ~/.conda and ~/.condarc, removed ~/miniconda3 and then unpacked everything and tried again. It is now working :-)

I've run all of the tests and get these test failures;

```
(MSMetaEnhancer) [opc@metawards MSMetaEnhancer]$ pytest tests/
=================================================================== test session starts ====================================================================
rootdir: /home/opc/MSMetaEnhancer
plugins: cov-3.0.0, dependency-0.5.1, aiohttp-0.3.0
collected 41 items

tests/test_BridgeDB.py ...                                                                                                                           [  7%]
tests/test_CIR.py ...                                                                                                                                [ 14%]
tests/test_CTS.py ....                                                                                                                               [ 24%]
tests/test_IDSM.py X..                                                                                                                               [ 31%]
tests/test_NLM.py ...                                                                                                                                [ 39%]
tests/test_PubChem.py .....                                                                                                                          [ 51%]
tests/test_annotator.py ....                                                                                                                         [ 60%]
tests/test_application.py ..                                                                                                                         [ 65%]
tests/test_converter.py .FFFFFF....                                                                                                                  [ 92%]
tests/test_curator.py ...                                                                                                                            [100%]

========================================================================= FAILURES =========================================================================
___________________________________________________ test_loop_request_circuit_breaker_get[TimeoutError] ____________________________________________________

self = <aiocircuitbreaker.aiocircuitbreaker.CircuitBreaker object at 0x7fa01e80ff10>, func = <function WebConverter.make_request at 0x7fa01e7d2040>
args = (<MSMetaEnhancer.libs.converters.web.WebConverter.WebConverter object at 0x7fa01ceb0f40>, '/', 'GET', None, None), kwargs = {}

    async def call(self, func: Callable[..., Any], *args: Any, **kwargs: Any) -> Any:
        """
        Calls the decorated function and applies the circuit breaker
        rules on success or failure
        :param func: Decorated async function
        """
        with self:
>           return await func(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:91:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <MSMetaEnhancer.libs.converters.web.WebConverter.WebConverter object at 0x7fa01ceb0f40>, url = '/', method = 'GET', data = None, headers = {}

    @circuit(failure_threshold=FAILURE_THRESHOLD,
             expected_exception=Union[TimeoutError, ServerDisconnectedError, ClientConnectorError],
             fallback_function=ServiceNotAvailable.raise_circuitbreaker)
    async def make_request(self, url, method, data, headers):
        """
        Enter a circuit breaker loop and execute request with type depending on specified method.

        :param url: converter URL
        :param method: GET/POST
        :param data: given arguments for POST request
        :param headers: optional headers for the request
        :return: obtained response
        """
        if headers is None:
            headers = dict()
        if method == 'GET':
>           async with self.session.get(url, headers=headers) as response:

MSMetaEnhancer/libs/converters/web/WebConverter.py:86:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656661344'>, args = ('/',), kwargs = {'headers': {}}

    def __call__(_mock_self, *args, **kwargs):
        # can't use self in-case a function / method we are mocking uses self
        # in the signature
        _mock_self._mock_check_sig(*args, **kwargs)
        _mock_self._increment_mock_call(*args, **kwargs)
>       return _mock_self._mock_call(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1100:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656661344'>, args = ('/',), kwargs = {'headers': {}}

    def _mock_call(_mock_self, *args, **kwargs):
>       return _mock_self._execute_mock_call(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1104:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656661344'>, args = ('/',), kwargs = {'headers': {}}, self = <Mock name='mock.get' id='140325656661344'>
effect = <class 'asyncio.exceptions.TimeoutError'>

    def _execute_mock_call(_mock_self, *args, **kwargs):
        self = _mock_self
        # separate from _increment_mock_call so that awaited functions are
        # executed separately from their call, also AsyncMock overrides this method

        effect = self.side_effect
        if effect is not None:
            if _is_exception(effect):
>               raise effect
E               asyncio.exceptions.TimeoutError

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1161: TimeoutError

failing_session_mock = <AsyncMock id='140325657133696'>

    async def test_loop_request_circuit_breaker_get(failing_session_mock):
        converter = WebConverter(failing_session_mock)

        with pytest.raises(ServiceNotAvailable):
>           await converter.loop_request('/', 'GET', None, None)

tests/test_converter.py:86:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
MSMetaEnhancer/libs/converters/web/WebConverter.py:105: in loop_request
    return await self.make_request(url, method, data, headers)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:77: in wrapper
    return await self.call(function, *args, **kwargs)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:91: in call
    return await func(*args, **kwargs)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:55: in __exit__
    if exc_type and issubclass(exc_type, self._expected_exception):  # type: ignore
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = typing.Union[asyncio.exceptions.TimeoutError, aiohttp.client_exceptions.ServerDisconnectedError, aiohttp.client_exceptions.ClientConnectorError]
cls = <class 'asyncio.exceptions.TimeoutError'>

    def __subclasscheck__(self, cls):
>       raise TypeError("Subscripted generics cannot be used with"
                        " class and instance checks")
E       TypeError: Subscripted generics cannot be used with class and instance checks

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/typing.py:723: TypeError
______________________________________________ test_loop_request_circuit_breaker_get[ServerDisconnectedError] ______________________________________________

self = <aiocircuitbreaker.aiocircuitbreaker.CircuitBreaker object at 0x7fa01e80ff10>, func = <function WebConverter.make_request at 0x7fa01e7d2040>
args = (<MSMetaEnhancer.libs.converters.web.WebConverter.WebConverter object at 0x7fa01ce2a730>, '/', 'GET', None, None), kwargs = {}

    async def call(self, func: Callable[..., Any], *args: Any, **kwargs: Any) -> Any:
        """
        Calls the decorated function and applies the circuit breaker
        rules on success or failure
        :param func: Decorated async function
        """
        with self:
>           return await func(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:91:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <MSMetaEnhancer.libs.converters.web.WebConverter.WebConverter object at 0x7fa01ce2a730>, url = '/', method = 'GET', data = None, headers = {}

    @circuit(failure_threshold=FAILURE_THRESHOLD,
             expected_exception=Union[TimeoutError, ServerDisconnectedError, ClientConnectorError],
             fallback_function=ServiceNotAvailable.raise_circuitbreaker)
    async def make_request(self, url, method, data, headers):
        """
        Enter a circuit breaker loop and execute request with type depending on specified method.

        :param url: converter URL
        :param method: GET/POST
        :param data: given arguments for POST request
        :param headers: optional headers for the request
        :return: obtained response
        """
        if headers is None:
            headers = dict()
        if method == 'GET':
>           async with self.session.get(url, headers=headers) as response:

MSMetaEnhancer/libs/converters/web/WebConverter.py:86:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656110272'>, args = ('/',), kwargs = {'headers': {}}

    def __call__(_mock_self, *args, **kwargs):
        # can't use self in-case a function / method we are mocking uses self
        # in the signature
        _mock_self._mock_check_sig(*args, **kwargs)
        _mock_self._increment_mock_call(*args, **kwargs)
>       return _mock_self._mock_call(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1100:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656110272'>, args = ('/',), kwargs = {'headers': {}}

    def _mock_call(_mock_self, *args, **kwargs):
>       return _mock_self._execute_mock_call(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1104:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656110272'>, args = ('/',), kwargs = {'headers': {}}, self = <Mock name='mock.get' id='140325656110272'>
effect = <class 'aiohttp.client_exceptions.ServerDisconnectedError'>

    def _execute_mock_call(_mock_self, *args, **kwargs):
        self = _mock_self
        # separate from _increment_mock_call so that awaited functions are
        # executed separately from their call, also AsyncMock overrides this method

        effect = self.side_effect
        if effect is not None:
            if _is_exception(effect):
>               raise effect
E               aiohttp.client_exceptions.ServerDisconnectedError: Server disconnected

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1161: ServerDisconnectedError

failing_session_mock = <AsyncMock id='140325655714448'>

    async def test_loop_request_circuit_breaker_get(failing_session_mock):
        converter = WebConverter(failing_session_mock)

        with pytest.raises(ServiceNotAvailable):
>           await converter.loop_request('/', 'GET', None, None)

tests/test_converter.py:86:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
MSMetaEnhancer/libs/converters/web/WebConverter.py:105: in loop_request
    return await self.make_request(url, method, data, headers)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:77: in wrapper
    return await self.call(function, *args, **kwargs)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:91: in call
    return await func(*args, **kwargs)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:55: in __exit__
    if exc_type and issubclass(exc_type, self._expected_exception):  # type: ignore
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = typing.Union[asyncio.exceptions.TimeoutError, aiohttp.client_exceptions.ServerDisconnectedError, aiohttp.client_exceptions.ClientConnectorError]
cls = <class 'aiohttp.client_exceptions.ServerDisconnectedError'>

    def __subclasscheck__(self, cls):
>       raise TypeError("Subscripted generics cannot be used with"
                        " class and instance checks")
E       TypeError: Subscripted generics cannot be used with class and instance checks

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/typing.py:723: TypeError
_______________________________________________ test_loop_request_circuit_breaker_get[failing_session_mock2] _______________________________________________

self = <aiocircuitbreaker.aiocircuitbreaker.CircuitBreaker object at 0x7fa01e80ff10>, func = <function WebConverter.make_request at 0x7fa01e7d2040>
args = (<MSMetaEnhancer.libs.converters.web.WebConverter.WebConverter object at 0x7fa01ce4ae20>, '/', 'GET', None, None), kwargs = {}

    async def call(self, func: Callable[..., Any], *args: Any, **kwargs: Any) -> Any:
        """
        Calls the decorated function and applies the circuit breaker
        rules on success or failure
        :param func: Decorated async function
        """
        with self:
>           return await func(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:91:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <MSMetaEnhancer.libs.converters.web.WebConverter.WebConverter object at 0x7fa01ce4ae20>, url = '/', method = 'GET', data = None, headers = {}

    @circuit(failure_threshold=FAILURE_THRESHOLD,
             expected_exception=Union[TimeoutError, ServerDisconnectedError, ClientConnectorError],
             fallback_function=ServiceNotAvailable.raise_circuitbreaker)
    async def make_request(self, url, method, data, headers):
        """
        Enter a circuit breaker loop and execute request with type depending on specified method.

        :param url: converter URL
        :param method: GET/POST
        :param data: given arguments for POST request
        :param headers: optional headers for the request
        :return: obtained response
        """
        if headers is None:
            headers = dict()
        if method == 'GET':
>           async with self.session.get(url, headers=headers) as response:

MSMetaEnhancer/libs/converters/web/WebConverter.py:86:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656242256'>, args = ('/',), kwargs = {'headers': {}}

    def __call__(_mock_self, *args, **kwargs):
        # can't use self in-case a function / method we are mocking uses self
        # in the signature
        _mock_self._mock_check_sig(*args, **kwargs)
        _mock_self._increment_mock_call(*args, **kwargs)
>       return _mock_self._mock_call(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1100:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656242256'>, args = ('/',), kwargs = {'headers': {}}

    def _mock_call(_mock_self, *args, **kwargs):
>       return _mock_self._execute_mock_call(*args, **kwargs)

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1104:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

_mock_self = <Mock name='mock.get' id='140325656242256'>, args = ('/',), kwargs = {'headers': {}}, self = <Mock name='mock.get' id='140325656242256'>
effect = ClientConnectorError(None, OSError())

    def _execute_mock_call(_mock_self, *args, **kwargs):
        self = _mock_self
        # separate from _increment_mock_call so that awaited functions are
        # executed separately from their call, also AsyncMock overrides this method

        effect = self.side_effect
        if effect is not None:
            if _is_exception(effect):
>               raise effect
E               aiohttp.client_exceptions.ClientConnectorError: <unprintable ClientConnectorError object>

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/mock/mock.py:1161: ClientConnectorError

failing_session_mock = <AsyncMock id='140325657300320'>

    async def test_loop_request_circuit_breaker_get(failing_session_mock):
        converter = WebConverter(failing_session_mock)

        with pytest.raises(ServiceNotAvailable):
>           await converter.loop_request('/', 'GET', None, None)

tests/test_converter.py:86:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
MSMetaEnhancer/libs/converters/web/WebConverter.py:105: in loop_request
    return await self.make_request(url, method, data, headers)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:77: in wrapper
    return await self.call(function, *args, **kwargs)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:91: in call
    return await func(*args, **kwargs)
../miniconda3/envs/MSMetaEnhancer/lib/python3.9/site-packages/aiocircuitbreaker/aiocircuitbreaker.py:55: in __exit__
    if exc_type and issubclass(exc_type, self._expected_exception):  # type: ignore
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = typing.Union[asyncio.exceptions.TimeoutError, aiohttp.client_exceptions.ServerDisconnectedError, aiohttp.client_exceptions.ClientConnectorError]
cls = <class 'aiohttp.client_exceptions.ClientConnectorError'>

    def __subclasscheck__(self, cls):
>       raise TypeError("Subscripted generics cannot be used with"
                        " class and instance checks")
E       TypeError: Subscripted generics cannot be used with class and instance checks

../miniconda3/envs/MSMetaEnhancer/lib/python3.9/typing.py:723: TypeError
[…]
