### 2022-09-19 9ce9cd7
Revert "Add pytest.fixture decorator to async test functions (required by newer pytest-aiohttp package)"
This reverts commit 0efde23e68034bca09d90cd581b127d75d1f85ba.
