### 2022-09-19 2bd9597
fix `test_lru_cache`
