### 2022-09-19 92aef6f
fix `test_loop_request`
