### 2022-09-19 8ae4232
fix `test_loop_request_fail`
