+++ tests/test_converter.py
@@ -52,15 +52,13 @@ async def test_loop_request(aiohttp_client):
 
 
 async def test_loop_request_fail(aiohttp_client):
-    async def fake_request(request):
+    async def fake_request():
         raise ServerDisconnectedError()
 
-    def create_app(loop):
-        app = web.Application(loop=loop)
-        app.router.add_route('GET', '/', fake_request)
-        return app
+    app = web.Application()
+    app.router.add_route('GET', '/', fake_request)
 
-    session = await aiohttp_client(create_app)
+    session = await aiohttp_client(app)
     converter = WebConverter(session)
 
     with pytest.raises(UnknownResponse):
