### 2022-09-20 7871bbd
make generics work with `isinstance()` in python 3.9
