### 2022-09-21 #4494 @chryswoods
I am away on extended leave and won't be able to look at this again until November. I am happy for the paper to be published once all of the unit tests pass.
