### 2022-09-28 #4494 @hechth
@csoneson @chryswoods I re-ran all tests manually and they all pass (and are actually run), as well as on the CI - see [here](https://github.com/RECETOX/MSMetaEnhancer/actions/runs/3144614358).

See the log output beneath for more information:

```
Running tests (pytest): c:\Users\473355\git\recetox\MSMetaEnhancer
Running test with arguments: --rootdir c:\Users\473355\git\recetox\MSMetaEnhancer --override-ini junit_family=xunit1 --junit-xml=C:\Temp\tmp-19804OKRM7VSNOeyQ.xml
Current working directory: c:\Users\473355\git\recetox\MSMetaEnhancer
Workspace directory: c:\Users\473355\git\recetox\MSMetaEnhancer
Run completed, parsing output
./tests/test_BridgeDB.py::test_service_available Passed

./tests/test_BridgeDB.py::test_format Passed

./tests/test_BridgeDB.py::test_get_conversions Passed

./tests/test_CIR.py::test_service_available Passed

./tests/test_CIR.py::test_format Passed

./tests/test_CIR.py::test_get_conversions Passed

./tests/test_CTS.py::test_format[7783-89-3-1] Passed

./tests/test_CTS.py::test_format[7783893-0] Passed

./tests/test_CTS.py::test_service_available Passed

./tests/test_CTS.py::test_get_conversions Passed

./tests/test_IDSM.py::test_service_available Passed

./tests/test_IDSM.py::test_format Passed

./tests/test_IDSM.py::test_get_conversions Passed

./tests/test_NLM.py::test_service_available Passed

./tests/test_NLM.py::test_format Passed

./tests/test_NLM.py::test_get_conversions Passed

./tests/test_PubChem.py::test_parse_attributes[response0-expected0] Passed

./tests/test_PubChem.py::test_parse_attributes[response1-expected1] Passed

./tests/test_PubChem.py::test_service_available Passed

./tests/test_PubChem.py::test_format Passed

./tests/test_PubChem.py::test_get_conversions Passed

./tests/test_annotator.py::test_annotate[data0-expected0-False-mocked0] Passed

./tests/test_annotator.py::test_annotate[data1-expected1-True-mocked1] Passed

./tests/test_annotator.py::test_execute_job_with_cache Passed

./tests/test_annotator.py::test_catch_exception Passed

./tests/test_application.py::test_annotate_spectra_monitor_stops Passed

./tests/test_application.py::test_annotate_spectra_monitor_stops_after_exception Passed

./tests/test_converter.py::test_loop_request_circuit_breaker_get[TimeoutError] Passed

./tests/test_converter.py::test_loop_request_circuit_breaker_get[ServerDisconnectedError] Passed

./tests/test_converter.py::test_loop_request_circuit_breaker_get[failing_session_mock2] Passed

./tests/test_converter.py::test_loop_request_circuit_breaker_post[TimeoutError] Passed

./tests/test_converter.py::test_loop_request_circuit_breaker_post[ServerDisconnectedError] Passed

./tests/test_converter.py::test_loop_request_circuit_breaker_post[failing_session_mock2] Passed

./tests/test_converter.py::test_process_request_exception[False-500] Passed

./tests/test_converter.py::test_process_request_exception[False-503] Passed

./tests/test_converter.py::test_query_the_service Passed

./tests/test_converter.py::test_loop_request Passed

./tests/test_converter.py::test_loop_request_fail Passed

./tests/test_converter.py::test_process_request Passed

./tests/test_converter.py::test_convert Passed

./tests/test_converter.py::test_lru_cache Passed

./tests/test_curator.py::test_filter_invalid_metadata[metadata0-validated_metadata0-1] Passed

./tests/test_curator.py::test_filter_invalid_metadata[metadata1-validated_metadata1-0] Passed

./tests/test_curator.py::test_fix_cas_number Passed

Total number of tests expected to run: 44
Total number of tests run: 44
Total number of tests passed: 44
Total number of tests failed: 0
Total number of tests failed with errors: 0
Total number of tests skipped: 0
Total number of tests with no result data: 0
Finished running tests!

> Test run finished at 28/09/2022, 17:28:38 <
```
