### 2022-10-04 35d6d39
bump MSME version to `0.2.4`
