### 2022-10-04 3bc4d37
bump MSME version to `0.2.4`
