+++ conda/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "MSMetaEnhancer" %}
-{% set version = "0.2.3" %}
+{% set version = "0.2.4" %}
 
 package:
   name: {{ name|lower }}
