### 2022-10-07 #4494 @csoneson
👋🏻 @hechth - apologies for the delay. I checked the tests and they look good to me. Hence, based on @chryswoods [comment](https://github.com/openjournals/joss-reviews/issues/4494#issuecomment-1253833103), I have checked the remaining boxes in the review checklist, and will proceed with the next steps. I will take a look through the submission, and get back to you shortly.
