### 2022-10-07 14ed347
Minor fixes to bibliography
