+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{yones2021metafetcher,
-  title={MetaFetcheR: An R package for complete mapping of small compound data},
+  title={MetaFetcheR: An {R} package for complete mapping of small compound data},
   author={Yones, Sara and Csombordi, Rajmund and Komorowski, Jan and Diamanti, Klev},
   journal={bioRxiv},
   year={2021},
@@ -8,14 +8,14 @@
 }
 
 @article{Huber2020,
-author = {Huber, Florian and Verhoeven, Stefan and Meijer, Christiaan and Spreeuw, Hanno and Castilla, Efra{\'{i}}n and Geng, Cunliang and van der Hooft, Justin and Rogers, Simon and Belloum, Adam and Diblen, Faruk and Spaaks, Jurriaan},
+author = {Huber, Florian and Verhoeven, Stefan and Meijer, Christiaan and Spreeuw, Hanno and Castilla, Efra{\'{i}}n and Geng, Cunliang and {van der Hooft}, Justin and Rogers, Simon and Belloum, Adam and Diblen, Faruk and Spaaks, Jurriaan},
 doi = {10.21105/joss.02411},
 issn = {2475-9066},
 journal = {Journal of Open Source Software},
 month = {aug},
 number = {52},
 pages = {2411},
-title = {matchms - processing and similarity evaluation of mass spectrometry data.},
+title = {{matchms - processing and similarity evaluation of mass spectrometry data.}},
 url = {https://joss.theoj.org/papers/10.21105/joss.02411},
 volume = {5},
 year = {2020}
@@ -69,7 +69,7 @@ year = {2020}
 }
 
 @article{cao2008chemminer,
-  title={ChemmineR: a compound mining framework for R},
+  title={ChemmineR: a compound mining framework for {R}},
   author={Cao, Yiqun and Charisi, Anna and Cheng, Li-Chang and Jiang, Tao and Girke, Thomas},
   journal={Bioinformatics},
   volume={24},
@@ -89,12 +89,13 @@ year = {2020}
 }
 
 @article{szocs2020webchem,
-  title={Webchem: an r package to retrieve chemical information from the Web},
+  title={Webchem: an {R} package to retrieve chemical information from the Web},
   author={Sz{\"o}cs, Eduard and Stirling, Tam{\'a}s and Scott, Eric R and Scharm{\"u}ller, Andreas and Sch{\"a}fer, Ralf B},
   journal={Journal of Statistical Software},
   volume={93},
   number={1},
   pages={1--17},
+  doi={10.18637/jss.v093.i13},
   year={2020}
 }
 
@@ -124,8 +125,8 @@ year = {2020}
 
 @article{galaxy,
   title = {{The {Galaxy} Platform for Accessible, Reproducible and Collaborative Biomedical Analyses: 2018 update}},
-  author = {Afgan, Enis and Baker, Dannon and Batut, B\'{e}r\'{e}nice and van den Beek, Marius and Bouvier, Dave and \v{C}ech, Martin and Chilton, John and Clements, Dave and Coraor, Nate and Gr\"uning, Bj\"orn A. and Guerler, Aysam and Hillman-Jackson, Jennifer and Hiltemann, Saskia and Jalili, Vahid and Rasche, Helena and Soranzo, Nicola and Goecks, Jeremy and Taylor, James and Nekrutenko, Anton and Blankenberg, Daniel},
-  journal = {Nucleic Acids Res.},
[…]
