### 2022-10-07 #4494 @csoneson
@hechth Alright, I went through the submission as well and overall it looks great - I opened a PR with a couple of minor fixes to the bibliography: https://github.com/RECETOX/MSMetaEnhancer/pull/128
Please take a look and merge if you agree. Is there a way to provide a DOI/URL for `RDKit`?

The next steps then are as follows:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
