### 2022-10-11 30ae42f
updated rdkit reference
