+++ paper/paper.bib
@@ -297,7 +297,41 @@ year = {2013}
 }
 
 @article{landrum2006rdkit,
-  title={RDKit: Open-source cheminformatics},
-  author={Landrum, Greg and others},
-  year={2006}
+  author     = {Greg Landrum and
+                  Paolo Tosco and
+                  Brian Kelley and
+                  Ric and
+                  sriniker and
+                  gedeck and
+                  Riccardo Vianello and
+                  NadineSchneider and
+                  Eisuke Kawashima and
+                  David Cosgrove and
+                  Andrew Dalke and
+                  Dan N and
+                  Gareth Jones and
+                  Brian Cole and
+                  Matt Swain and
+                  Samo Turk and
+                  AlexanderSavelyev and
+                  Alain Vaucher and
+                  Maciej Wójcikowski and
+                  Ichiru Take and
+                  Daniel Probst and
+                  Kazuya Ujihara and
+                  Vincent F. Scalfani and
+                  guillaume godin and
+                  Axel Pahl and
+                  Francois Berenger and
+                  JLVarjo and
+                  strets123 and
+                  JP and
+                  DoliathGavid},
+  title        = {rdkit/rdkit: 2022\_03\_5 (Q1 2022) Release},
+  month        = aug,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {Release\_2022\_03\_5},
+  doi          = {10.5281/zenodo.6961488},
+  url          = {https://doi.org/10.5281/zenodo.6961488}
 }
