### 2022-10-15 #4494 @hechth
@csoneson I made a new release 0.2.5 which is archived on Zenodo with DOI: 10.5281/zenodo.7211726
 at https://zenodo.org/record/7211726
