### 2022-10-22 #4494 @egonw
@hechth, @csoneson, not sure about the exact status, but BridgeDb is with a lower case 'b' at the end. Can still this be updated? If not, it's a very minor thing.
