### 2022-10-24 a573b2e
Renamed BridgeDB class to BridgeDb
