+++ tests/test_BridgeDB.py
@@ -1,7 +1,7 @@
 import asyncio
 import pytest
 
-from MSMetaEnhancer.libs.converters.web import BridgeDB
+from MSMetaEnhancer.libs.converters.web import BridgeDb
 from tests.utils import wrap_with_session
 
 
@@ -10,13 +10,13 @@ HMDBID = 'HMDB0000001'
 
 @pytest.mark.dependency()
 def test_service_available():
-    asyncio.run(wrap_with_session(BridgeDB, 'hmdbid_to_pubchemid', ['HMDB0000001']))
+    asyncio.run(wrap_with_session(BridgeDb, 'hmdbid_to_pubchemid', ['HMDB0000001']))
 
 
 @pytest.mark.dependency(depends=["test_service_available"])
 def test_format():
     args = f'Ch/{HMDBID}'
-    response = asyncio.run(wrap_with_session(BridgeDB, 'query_the_service', ['BridgeDB', args]))
+    response = asyncio.run(wrap_with_session(BridgeDb, 'query_the_service', ['BridgeDB', args]))
 
     assert type(response) == str
     lines = response.split('\n')
@@ -25,5 +25,5 @@ def test_format():
 
 
 def test_get_conversions():
-    jobs = BridgeDB(None).get_conversion_functions()
+    jobs = BridgeDb(None).get_conversion_functions()
     assert ('wikidataid', 'pubchemid', 'BridgeDB') in jobs
