### 2022-10-24 #4494 @hechth
@csoneson @egonw  Should I then also draft a new release with the code changes, or is it okay to only have the changes in the paper for now and just publish the fixes in the next code release?
