+++ MSMetaEnhancer/libs/converters/web/__init__.py
@@ -3,6 +3,6 @@ from MSMetaEnhancer.libs.converters.web.CTS import CTS
 from MSMetaEnhancer.libs.converters.web.CIR import CIR
 from MSMetaEnhancer.libs.converters.web.NLM import NLM
 from MSMetaEnhancer.libs.converters.web.PubChem import PubChem
-from MSMetaEnhancer.libs.converters.web.BridgeDb import BridgeDb
+from MSMetaEnhancer.libs.converters.web.BridgeDB import BridgeDb
 
 __all__ = ['IDSM', 'CTS', 'CIR', 'NLM', 'PubChem', 'BridgeDb']
