### 2022-10-24 #4494 @egonw
I'd be more than happy with it comes with the next code release.
