### 2022-11-01 #4494 @hechth
@csoneson are there any news from the EIC regarding the paper acceptance?
