### 2022-11-02 #4494 @Kevin-Mattheus-Moerman
@hechth I have read the paper and reviewed the ZENODO archive and all seems in order. I only have one minor comment:

- [ ] Your ZENODO archive lists a Creative Commons license. However the software carries an MIT license. We can proceed here with acceptance but please do manually edit the ZENODO archive so the license entry to matches your software's license.
