### 2022-11-02 #4494 @hechth
@Kevin-Mattheus-Moerman I changed the license on Zenodo - I never tried actually typing a license in there since the dropdown only lists the creative commons licenses - now I changed it to MIT :).
