### 2022-11-02 #4494 @Kevin-Mattheus-Moerman
Congratulations @hechth on getting this work published in JOSS! :tada:

Thanks for editing this work @csoneson.

Also special thanks to @marshallmcdonnell and @chryswoods for reviewing!!!! :rocket:
